Zero functionality now functioning
The zeroing function now works. It moves up 2cm and then right 2cm first (in case it was already zeroed), and then moves at 1/100th the speed of normal operation. Moves 1 step at a time along each axis until the zeroing switch is hit.

Have yet to test with the mounted switches on the board, but implementation is there.