Acceleration added to movements
If the movement is above a certain distance, the motors are accelerated from a minimum to a maximum speed over one rotation. This is to help prevent the motors from skipping sometimes as we have had issues with this. So far, the acceleration seems to have solved this with very minimal impact on the duration of movement. Max speed has also been increased from 100kHz to 200kHz (the maximum supported by the TB6600 drivers).