Rudimentary UI for the yolov7 model
Super rough GUI for the TensorRT model. Has a startup screen, and another that shows the inference.

I have to create the UI objects in the inference thread since TKinter cannot be run outside of main (see https://stackoverflow.com/a/45803955). To get around it I build the UI objects in the inference thread. TKinter still throws an error on exit, but by then the program has finished running so its "fine".

Next steps is to make the UI more advanced and see if it has any affect on the inference times.