Adjust performance calculation
Adjust the performance calculation again.

The normalization is still performed with the max and minimum values. The minimum however will be the lowest non-zero value.

Rationale was that a missed punch results in a zero (or near zero) score that skews the normalized values. This approach aims to give better looking results that stay true to the purpose of the metric.